+++ src/main/java/net/Mega2223/AguaEngine/graphics/objects/shapes/Image.java
@@ -1,9 +1,9 @@
 package net.Mega2223.AguaEngine.graphics.objects.shapes;
 
 import java.awt.image.BufferedImage;
-
+//todo
 public class Image extends Shape  {
-    public float[] size = {1,1};
+    public double[] size = {1,1};
     public BufferedImage image;
     public Image(BufferedImage image){
         this.image = image;
